# swagger.yaml
openapi: 3.0.3
info:
  title: Employee Management API
  description: |
    A serverless employee management system built with AWS Lambda, DynamoDB, API Gateway, and Cognito authentication.
    
    ## Authentication
    Most endpoints require authentication using AWS Cognito JWT tokens. Include the access token in the Authorization header:
    ```
    Authorization: Bearer {accessToken}
    ```
    
    ## Base URL
    - Development: https://ec4msgjtx3.execute-api.ap-south-1.amazonaws.com/dev
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: ISC
servers:
  - url: https://ec4msgjtx3.execute-api.ap-south-1.amazonaws.com/dev
    description: Development server
  - url: https://ec4msgjtx3.execute-api.ap-south-1.amazonaws.com/prod
    description: Production server
...
paths:
...
  /auth/confirm:
    post:
      tags:
        - Authentication
      summary: Confirm user registration
      description: Verify user email with the confirmation code sent during signup
      operationId: confirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - code
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                code:
                  type: string
                  example: "123456"
                  description: 6-digit confirmation code from email
      responses:
        '200':
          description: User confirmed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User confirmed successfully. You can now sign in.
        '400':
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT access token. Obtain by calling /auth/signin
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error description
        errors:
          type: array
          items:
            type: string
          example: ["Error detail 1", "Error detail 2"]
        code:
          type: string
          example: ErrorCode
...